+++ etc/ipython/profile_default/startup/30-settings.py
@@ -2,8 +2,7 @@
 # pylint: disable=unused-import
 # pylint: disable=wrong-import-position
 # pylint: disable=invalid-name
-"""
-IPython startup-file, outside of PYTHONPATH.
+"""IPython startup-file, outside of PYTHONPATH.
 
 Files in this startup-folder will be run in lexicographical order,
 so you can control the execution order of files with a prefix, e.g.::
@@ -18,7 +17,6 @@ return-statements are not allowed.
 
 print(f"Importing {__file__}")
 
-import libranet_flask
 from libranet_flask import cfg
 
 settings = cfg.get_settings()