+++ etc/pythonstartup.py
@@ -1,4 +1,3 @@
-# ruff: noqa: D100 - Missing docstring in public module
 """Python startup file.
 
 Taken from https://gist.github.com/viliampucik/8713b09ff7e4d984b29bfcd7804dc1f4