+++ src/libranet_flask/__init__.py
@@ -2,7 +2,11 @@
 
 from libranet_flask.about import (
     authors as __author__,
+)
+from libranet_flask.about import (
     license_ as __license__,
+)
+from libranet_flask.about import (
     version as __version__,
 )