+++ src/libranet_flask/cli/__init__.py
@@ -14,4 +14,3 @@ app.command(obj=app_subcmd1)
 @app.default
 def main() -> None:
     """Run the main command."""
-    pass