+++ src/libranet_flask/cli/flask.py
@@ -2,6 +2,7 @@
 
 Docs: https://flask.palletsprojects.com/en/2.2.x/cli/
 """
+
 import click
 import flask
 
@@ -11,9 +12,9 @@ demo = flask.Blueprint("demo", __name__)
 @demo.cli.command("hello")
 @click.option("--name", default="World")
 def hello(name: str) -> None:
-    """
+    """Say hello.
 
     Usage:
-        > bin/flask demo hello --name Foo
+    > bin/flask demo hello --name Foo.
     """
     click.echo(f"Hello, {name}!")