+++ src/libranet_flask/cli/subcmd1.py
@@ -10,8 +10,7 @@ app_subcmd1: cyclopts.App = cyclopts.App(
 
 @app_subcmd1.command(name="start", sort_key=10)
 def start_instances(instance: str) -> None:
-    """
-    Start instance.
+    """Start instance.
 
     Args:
         instance: instance to start
@@ -21,8 +20,7 @@ def start_instances(instance: str) -> None:
 
 @app_subcmd1.command(name="stop", sort_key=20)
 def stop_instances(instance: str) -> None:
-    """
-    Stop instance.
+    """Stop instance.
 
     Args:
         instance: instance to stop