+++ docs/changes.md
@@ -2,8 +2,6 @@
 
 All notable changes to this project will be documented in this file.
 
-
-
 ## 1.0.2 (2023-03-20)
 
-- Package created by [Wouter Vanden Hove <wouter@libranet.eu>]
+- Package created by \[Wouter Vanden Hove <wouter@libranet.eu>\]